Added functions package, fixed imports
