+++ napari_live_recording/Functions/__init__.py
@@ -1,0 +1,5 @@
+from .functions import acquire
+
+__all__ = [
+    "acquire"
+]
